Computing the block offsets instead of split offsets
This approach gives you a better visibility to the record boundary. We still need recordsRead field since the actual record offset is impossible to detect when the underlying sequence file is BLOCK compressed.
The current implementation will give you record offsets with recordsRead as 0 when we have RECORD compression / no compression on the sequence file