Fixing the path computation for Consolidator
Thanks to all the test cases in PailOps we now have a (hopefully) much faster split calculation for Consolidator. Since the entire consolidator is designed to run for files under the same vertical partition, we group files based on that and process the split accordingly.

We might want to watch out for potential OOM because of the Consolidator#groupByParentPaths. Earlier we used to get files only under a prefix and process them but now we process all the files under the instanceRoot and process them.